Graceful signal handler (1/2)
The signal handler is supposed to allow for a graceful shutdown, not
just exit immediately, which would happen without it anyway.

Wait for the current run to end and only then exit if a signal has been
received. This also requires specific configuration to the systemd unit
file, so that file is included as well. More specifically, KillMode
needs to be set to "process" (we'd get broken pipe otherwise) and since
the signal may be received at the start of a run, increase the timeout.