[phab2wiki sync] changed files list
There are different ways to identify the files that need to be updated
on Wikipedia. The simplest would've been to use the file timestamps, but
it wouldn't be overly reliable. Instead, implement a wrapper around git
pull that parses git diff-tree to obtain and keep a list of the changed
files. We'll use that list later.