Completely rework the Phab2Wiki sync
Instead of just syncing the pages to the last update on Phabricator, we
now process the commits from Phabricator one by one, converting them
into edits on the Wiki with the respective committer, commit date and
commit message put into the summary. This should allow for much better
understanding of why a certain change was made by the bot.

In the future, it might be interesting to investigate whether it would
be possible to include additional information into the summary, for
example links to the respective Differential revisions.