Handle properly newlines at EOF
Most pages on the wiki don't have \n at the end, as it would be rendered
as a superfluous empty line. But this isn't the case with files, where
the lack of \n at EOF of text files is generally to be avoided. Hence we
now add \n when we update the files and remove it when update pages.
This should also preserve the existing \n at the end of pages.