Allow commits to be ignored when syncing
This might introduce nasty out-of-sync conditions, but may be helpful in
some situations, e.g. when we need to fix the \n at EOL.

To notify the bot that certain commit must be ignored, put anywhere in
the commit message "DO NOT SYNC" or "DO NOT MERGE". Case is important,
but surrounding brackets are not (e.g. you may use "[DO NOT MERGE]").