Support forced re-sync and deleted pages on-wiki
It's possible that sometimes the wiki pages and our repos may get out of
sync (e.g. because of 995f1b50f630be5d81d946002477a0ba04965e1a). Allow
this to be fixed by forcing an update of all pages to the repo.

Also, detect deleted pages on the wiki and remove them from our repos.
This also includes moved pages, but because each action is translated
into a separate commit, git will not detect the move as rename.

Please note as well that renaming files in the repos on Phabricator is
undesirable: instead of moving the page on the wiki, we'll delete the
original and create the target anew, thus losing all the history.

We may add specific support for such actions in the future.

Lastly, some of the code has been refactored and minor bugs fixed.